requirements.txt
python-dotenv

# Elasticsearch 5.x
elasticsearch>=5.0.0,<6.0.0

PyGithub
